+++ components/scheduler/CHANGELOG.md
@@ -4,4 +4,6 @@
 
 ## New Features
 
+- [Scheduler] Added `onError` custom event when an error occurs [#262](https://github.com/nylas/components/pull/262)
+
 ## Bug Fixes